breaking dyson_magnus_solvers.py into 4 files
