+++ qiskit_dynamics/__init__.py
@@ -29,7 +29,8 @@ from .signals.signals import Signal, DiscreteSignal
 
 from .solvers.solver_functions import solve_ode, solve_lmde
 from .solvers.solver_classes import Solver
-from .solvers.dyson_magnus_solvers import DysonSolver, MagnusSolver
+from .solvers.perturbative_solvers.dyson_solver import DysonSolver
+from .solvers.perturbative_solvers.magnus_solver import MagnusSolver
 
 from . import models
 from . import signals
