+++ qiskit_dynamics/solvers/perturbative_solvers/expansion_model.py
@@ -17,13 +17,10 @@
 Model class for perturbative expansion of solution of an LMDE.
 """
 
-from abc import ABC, abstractmethod
 from typing import Optional, Tuple, Callable, List, Union
 
 import numpy as np
 from numpy.polynomial.chebyshev import chebpts1, chebvander, chebval
-from scipy.linalg import expm
-from scipy.integrate._ivp.ivp import OdeResult
 
 from multiset import Multiset
 
@@ -31,357 +28,19 @@ from qiskit import QiskitError
 from qiskit.quantum_info import Operator
 
 from qiskit_dynamics import Signal, RotatingFrame
-from qiskit_dynamics.signals import SignalList
 from qiskit_dynamics.perturbation import solve_lmde_perturbation, ArrayPolynomial
 from qiskit_dynamics.array import Array
 from qiskit_dynamics.type_utils import to_array
-from qiskit_dynamics.dispatch.dispatch import Dispatch
-
-from .solver_utils import setup_args_lists
 
 try:
     import jax.numpy as jnp
-    from jax import vmap
-    from jax.scipy.linalg import expm as jexpm
-    from jax.lax import scan, associative_scan
+    from jax.lax import scan
 except ImportError:
     pass
 
 
-class PerturbativeSolver(ABC):
-    """Abstract base class for perturbation-based solvers."""
-
-    @abstractmethod
-    def __init__(self, model: "ExpansionModel"):
-        r"""Initialize.
-
-        Args:
-            model: ExpansionModel defining solver behaviour.
-        """
-        self._model = model
-
-    @property
-    def model(self) -> "ExpansionModel":
-        """Model object storing expansion details."""
-        return self._model
-
-    def solve(
-        self,
-        t0: Union[float, List[float]],
-        n_steps: Union[int, List[int]],
-        y0: Union[Array, List[Array]],
-        signals: Union[List[Signal], List[List[Signal]]],
-    ) -> Union[OdeResult, List[OdeResult]]:
-        """Solve given an initial time, number of steps, signals, and initial state.
-
-        Note that this method can be used to solve a list of simulations at once, by specifying
-        one or more of the arguments ``t0``, ``n_steps``, ``y0``, or ``signals`` as a list of
-        valid inputs. For this mode of operation, all of these arguments must be either lists of
-        the same length, or a single valid input, which will be used repeatedly.
-
-        Args:
-            t0: Initial time.
-            n_steps: Number of time steps to solve for.
-            y0: Initial state at time t0.
-            signals: List of signals.
-
-        Returns:
-            OdeResult: Results object, or list of results objects.
-
-        Raises:
-            QiskitError: If improperly formatted arguments.
-        """
-
-        # validate and setup list of simulations
-        [t0_list, n_steps_list, y0_list, signals_list], multiple_sims = setup_args_lists(
-            args_list=[t0, n_steps, y0, signals],
-            args_names=["t0", "n_steps", "y0", "signals"],
-            args_to_list=[
-                lambda x: scalar_to_list(x, "t0"),
-                lambda x: scalar_to_list(x, "n_steps"),
-                y0_to_list,
-                signals_to_list,
-            ],
-        )
-
-        all_results = []
-        for args in zip(t0_list, n_steps_list, y0_list, signals_list):
-            if len(args[-1]) != len(self.model.operators):
-                raise QiskitError("Signals must be the same length as the operators in the model.")
-            all_results.append(
-                self._solve(t0=args[0], n_steps=args[1], y0=args[2], signals=args[3])
-            )
-
-        if multiple_sims is False:
-            return all_results[0]
-
-        return all_results
-
-    @abstractmethod
-    def _solve(self, t0: float, n_steps: int, y0: Array, signals: List[Signal]) -> OdeResult:
-        """Solve once for a list of signals, initial state, initial time, and number of steps.
-
-        Args:
-            t0: Initial time.
-            n_steps: Number of time steps to solve for.
-            y0: Initial state at time t0.
-            signals: List of signals.
-
-        Returns:
-            OdeResult: Results object.
-        """
-        pass
-
-
-class DysonSolver(PerturbativeSolver):
-    r"""Solver for linear matrix differential equations based on the Dyson series.
-
-    This class implements the Dyson-series based solver presented in
-    [:footcite:`puzzuoli_sensitivity_2022`], which is a variant of the *Dysolve* algorithm
-    originally introduced in [:footcite:p:`shillito_fast_2020`].
-    This solver applies to linear matrix differential equations with generators of the form:
-
-    .. math::
-
-        G(t) = G_0 + \sum_{j=1}^s \textnormal{Re}[f_j(t) e^{i 2 \pi \nu_j t}]G_j,
-
-    and solves the LMDE in the rotating frame of :math:`G_0`, which is assumed to be
-    anti-Hermitian. I.e. it solves the LMDE with generator:
-
-    .. math::
-
-        \tilde{G}(t) = \sum_{j=1}^s \textnormal{Re}[f_j(t) e^{i 2 \pi \nu_j t}]\tilde{G}_j(t),
-
-    with :math:`\tilde{G}_i(t) = e^{-t G_0} G_i e^{tG_0}`. The solver is *fixed-step*,
-    with step size :math:`\Delta t` being defined at instantiation,
-    and solves over each step by computing a truncated Dyson series. See the
-    :ref:`Time-dependent perturbation theory and multi-variable
-    series expansions review <perturbation review>` for a description of the Dyson series.
-
-    At instantiation, the following parameters, which define the structure of the Dyson series
-    used, are fixed:
-
-    - The step size :math:`\Delta t`,
-    - The operator structure :math:`G_0`, :math:`G_i`,
-    - The reference frequencies :math:`\nu_j`,
-    - Approximation schemes for the envelopes :math:`f_j` over each time step (see below), and
-    - The Dyson series terms to keep in the truncation.
-
-    A 'compilation' or 'pre-computation' step computing the truncated expansion occurs
-    at instantiation. Once instantiated, the LMDE can be solved repeatedly for different lists of
-    envelopes :math:`f_1(t), \dots, f_s(t)` by calling the :meth:`solve` method with the
-    initial time ``t0`` and number of time-steps ``n_steps`` of size :math:`\Delta t`.
-    The list of envelopes are specified as :class:`~qiskit_dynamics.signals.Signal` objects,
-    whose carrier frequencies are automatically shifted to the reference frequencies :math:`\nu_j`,
-    with the frequency difference, and any phase, being absorbed into the envelope.
-
-    More explicitly, the process of solving over an interval :math:`[t_0, t_0 + \Delta t]`
-    is as follows. After shifting the carrier frequencies, the resulting envelopes are approximated
-    using a discrete Chebyshev transform, whose orders for each signal is given by
-    ``chebyshev_orders``. That is, for :math:`t \in [t_0, t_0 + \Delta t]`, each envelope is
-    approximated as:
-
-    .. math::
-
-        f_j(t) \approx \sum_{m=0}^{d_j} f_{j,m}T_m(t-t_0)
-
-    where :math:`T_m(\cdot)` are the Chebyshev polynomials over the interval,
-    :math:`f_{j,m}` are the approximation coefficients attained via Discrete Chebyshev Transform,
-    and :math:`d_j` is the order of approximation used for the given envelope.
-    Using:
-
-    .. math::
-
-        \textnormal{Re}[f_{j,m}T_m(t-t_0)e^{i2 \pi \nu_j t}] =
-        \textnormal{Re}[f_{j,m}e^{i 2 \pi \nu_j t_0}] \cos(2 \pi \nu_j (t-t_0))T_m(t-t_0) \\
-        + \textnormal{Im}[f_{j,m}e^{i 2 \pi \nu_j t_0}] \sin(-2 \pi \nu_j (t-t_0))T_m(t-t_0)
-
-    The generator is approximately decomposed as
-
-    .. math::
-
-        \tilde{G}(t) \approx \sum_{j=1}^s \sum_{m=0}^{d_j}
-        \textnormal{Re}[f_{j,m}e^{i 2 \pi \nu_j t_0}]
-        \cos(2 \pi \nu_j (t-t_0))T_m(t-t_0) \tilde{G}_j \\
-        + \sum_{j=1}^s \sum_{m=0}^{d_j}
-        \textnormal{Im}[f_{j,m}e^{i 2 \pi \nu_j t_0}]
-        \sin(- 2 \pi \nu_j (t-t_0))T_m(t-t_0) \tilde{G}_j
-
-    The multivariable Dyson series is then computed relative to the
-    above decomposition, with the variables being the
-    :math:`\textnormal{Re}[f_{j,m}e^{i 2 \pi \nu_j t_0}]` and
-    :math:`\textnormal{Im}[f_{j,m}e^{i 2 \pi \nu_j t_0}]`, and the operators being
-    :math:`\cos(2 \pi \nu_j (t-t_0))T_m(t-t_0) G_j` and
-    :math:`\sin(- 2 \pi \nu_j (t-t_0))T_m(t-t_0) G_j`. As shown in
-    [:footcite:`puzzuoli_sensitivity_2022`, :footcite:p:`shillito_fast_2020`],
-    the multivariable Dyson series for intervals of length :math:`\Delta t`
-    with different starting times are related via a simple frame change, and as such
-    these need only be computed once, and this makes up the 'pre-computation' step of this
-    object.
-
-    The optional argument ``include_imag`` can be used to control, on a signal by signal basis,
-    whether or not the imaginary terms
-
-    .. math::
-
-        \textnormal{Im}[f_{j,m}e^{i 2 \pi \nu_j t_0}]
-        \sin(- 2 \pi \nu_j (t-t_0))T_m(t-t_0) \tilde{G}_j
-
-    are included in the scheme. In generality they are required, but in special cases they are not
-    necessary, such as when :math:`\nu_j = 0`, or if :math:`f_j(t)`, including phase, is purely
-    real. By default all such terms are included.
-
-    .. footbibliography::
-    """
-
-    def __init__(
-        self,
-        operators: List[Operator],
-        rotating_frame: Union[Array, Operator, RotatingFrame, None],
-        dt: float,
-        carrier_freqs: Array,
-        chebyshev_orders: List[int],
-        expansion_order: Optional[int] = None,
-        expansion_labels: Optional[List[Multiset]] = None,
-        integration_method: Optional[str] = None,
-        include_imag: Optional[List[bool]] = None,
-        **kwargs,
-    ):
-        r"""Initialize.
-
-        Args:
-            operators: List of constant operators specifying the operators with signal coefficients.
-            rotating_frame: Rotating frame to setup the solver in.
-                            Must be Hermitian or anti-Hermitian.
-            dt: Fixed step size to compile to.
-            carrier_freqs: Carrier frequencies of the signals in the generator decomposition.
-            chebyshev_orders: Approximation degrees for each signal over the interval [0, dt].
-            expansion_order: Order of perturbation terms to compute up to. Specifying this
-                             argument results in computation of all terms up to the given order.
-                             Can be used in conjunction with ``expansion_terms``.
-            expansion_labels: Specific perturbation terms to compute. If both ``expansion_order``
-                              and ``expansion_terms`` are specified, then all terms up to
-                              ``expansion_order`` are computed, along with the additional terms
-                              specified in ``expansion_terms``. Labels are specified either as
-                              ``Multiset`` or as valid arguments to the ``Multiset`` constructor.
-                              This function further requires that ``Multiset``\s consist only of
-                              non-negative integers.
-            integration_method: ODE solver method to use when computing perturbation terms.
-            include_imag: List of bools determining whether to keep imaginary components in
-                          the signal approximation. Defaults to True for all signals.
-            kwargs: Additional arguments to pass to the solver when computing perturbation terms.
-        """
-        model = ExpansionModel(
-            operators=operators,
-            rotating_frame=rotating_frame,
-            dt=dt,
-            carrier_freqs=carrier_freqs,
-            chebyshev_orders=chebyshev_orders,
-            expansion_method="dyson",
-            expansion_order=expansion_order,
-            expansion_labels=expansion_labels,
-            integration_method=integration_method,
-            include_imag=include_imag,
-            **kwargs,
-        )
-        super().__init__(model=model)
-
-    def _solve(self, t0: float, n_steps: int, y0: Array, signals: List[Signal]) -> OdeResult:
-        ys = None
-        if Array.default_backend() == "jax":
-            single_step = lambda x: self.model.evaluate(x).data
-            ys = [y0, perturbative_solve_jax(single_step, self.model, signals, y0, t0, n_steps)]
-        else:
-            single_step = lambda coeffs, y: self.model.evaluate(coeffs) @ y
-            ys = [y0, perturbative_solve(single_step, self.model, signals, y0, t0, n_steps)]
-
-        return OdeResult(t=[t0, t0 + n_steps * self.model.dt], y=ys)
-
-
-class MagnusSolver(PerturbativeSolver):
-    """Solver for linear matrix differential equations based on the Magnus expansion.
-
-    This class implements the Magnus expansion-based solver presented in
-    [:footcite:`puzzuoli_sensitivity_2022`], which is a Magnus expansion variant of the
-    *Dysolve* algorithm originally introduced in [:footcite:p:`shillito_fast_2020`]. Its
-    setup and behaviour are the same as as the :class:`~qiskit_dynamics.solvers.DysonSolver`
-    class, with the sole exception being that it uses a truncated Magnus expansion
-    and matrix exponentiation to solve over a single time step. See the
-    :ref:`Time-dependent perturbation theory and multi-variable
-    series expansions review <perturbation review>` for a description of the Magnus expansion,
-    and the documentation for :class:`~qiskit_dynamics.solvers.DysonSolver` for more detailed
-    behaviour of this class.
-    """
-
-    def __init__(
-        self,
-        operators: List[Operator],
-        rotating_frame: Union[Array, Operator, RotatingFrame, None],
-        dt: float,
-        carrier_freqs: Array,
-        chebyshev_orders: List[int],
-        expansion_order: Optional[int] = None,
-        expansion_labels: Optional[List[Multiset]] = None,
-        integration_method: Optional[str] = None,
-        include_imag: Optional[List[bool]] = None,
-        **kwargs,
-    ):
-        r"""Initialize.
-
-        Args:
-            operators: List of constant operators specifying the operators with signal coefficients.
-            rotating_frame: Rotating frame to setup the solver in.
-                            Must be Hermitian or anti-Hermitian.
-            dt: Fixed step size to compile to.
-            carrier_freqs: Carrier frequencies of the signals in the generator decomposition.
-            chebyshev_orders: Approximation degrees for each signal over the interval [0, dt].
-            expansion_order: Order of perturbation terms to compute up to. Specifying this
-                             argument results in computation of all terms up to the given order.
-                             Can be used in conjunction with ``expansion_terms``.
-            expansion_labels: Specific perturbation terms to compute. If both ``expansion_order``
-                              and ``expansion_terms`` are specified, then all terms up to
-                              ``expansion_order`` are computed, along with the additional terms
-                              specified in ``expansion_terms``. Labels are specified either as
-                              ``Multiset`` or as valid arguments to the ``Multiset`` constructor.
-                              This function further requires that ``Multiset``\s consist only of
-                              non-negative integers.
-            integration_method: ODE solver method to use when computing perturbation terms.
-            include_imag: List of bools determining whether to keep imaginary components in
-                          the signal approximation. Defaults to True for all signals.
-            kwargs: Additional arguments to pass to the solver when computing perturbation terms.
-        """
-        model = ExpansionModel(
-            operators=operators,
-            rotating_frame=rotating_frame,
-            dt=dt,
-            carrier_freqs=carrier_freqs,
-            chebyshev_orders=chebyshev_orders,
-            expansion_method="magnus",
-            expansion_order=expansion_order,
-            expansion_labels=expansion_labels,
-            integration_method=integration_method,
-            include_imag=include_imag,
-            **kwargs,
-        )
-        super().__init__(model=model)
-
-    def _solve(self, t0: float, n_steps: int, y0: Array, signals: List[Signal]) -> OdeResult:
-        ys = None
-        if Array.default_backend() == "jax":
-            single_step = lambda x: self.model.Udt @ jexpm(self.model.evaluate(x).data)
-            ys = [y0, perturbative_solve_jax(single_step, self.model, signals, y0, t0, n_steps)]
-        else:
-            single_step = lambda coeffs, y: self.model.Udt @ expm(self.model.evaluate(coeffs)) @ y
-            ys = [y0, perturbative_solve(single_step, self.model, signals, y0, t0, n_steps)]
-
-        return OdeResult(t=[t0, t0 + n_steps * self.model.dt], y=ys)
-
-
 class ExpansionModel:
-    """Represents a perturbative expansion of an LMDE.
-
-    Should this just straight up do the construction?
-    """
+    """Represents a perturbative expansion of an LMDE."""
 
     def __init__(
         self,
@@ -549,55 +208,6 @@ class ExpansionModel:
     def evaluate(self, coeffs: Array) -> Array:
         """Evaluate the expansion polynomial for a given list of coefficients."""
         return self.expansion_polynomial(coeffs)
-
-
-def perturbative_solve(
-    single_step: Callable,
-    model: "ExpansionModel",
-    signals: List[Signal],
-    y0: np.ndarray,
-    t0: float,
-    n_steps: int,
-) -> np.ndarray:
-    """Standard python logic version of perturbative solving routine."""
-    U0 = model.rotating_frame.state_out_of_frame(t0, np.eye(model.Udt.shape[0], dtype=complex))
-    Uf = model.rotating_frame.state_into_frame(
-        t0 + n_steps * model.dt, np.eye(U0.shape[0], dtype=complex)
-    )
-
-    sig_cheb_coeffs = model.approximate_signals(signals, t0, n_steps)
-
-    y = U0 @ y0
-    for k in range(n_steps):
-        y = single_step(sig_cheb_coeffs[:, k], y)
-
-    return Uf @ y
-
-
-def perturbative_solve_jax(
-    single_step: Callable,
-    model: "ExpansionModel",
-    signals: List[Signal],
-    y0: np.ndarray,
-    t0: float,
-    n_steps: int,
-) -> np.ndarray:
-    """JAX version of perturbative_solve."""
-    U0 = model.rotating_frame.state_out_of_frame(
-        t0, jnp.eye(model.Udt.shape[0], dtype=complex)
-    ).data
-    Uf = model.rotating_frame.state_into_frame(
-        t0 + n_steps * model.dt, jnp.eye(U0.shape[0], dtype=complex)
-    ).data
-
-    sig_cheb_coeffs = model.approximate_signals(signals, t0, n_steps).data
-
-    y = U0 @ y0
-
-    step_propagators = vmap(single_step)(jnp.flip(sig_cheb_coeffs.transpose(), axis=0))
-    y = associative_scan(jnp.matmul, step_propagators, axis=0)[-1] @ y
-
-    return Uf @ y
 
 
 def construct_cheb_perturbations(
@@ -935,61 +545,3 @@ def construct_DCT(degree: int, domain: Optional[List] = None) -> Tuple:
     dct_mat[1:] /= 0.5 * order
 
     return dct_mat, xcheb_shifted
-
-
-def scalar_to_list(x, name):
-    """Check if x is a scalar or a list of scalars, and convert to a list in either case."""
-    was_list = False
-    x_ndim = nested_ndim(x)
-    if x_ndim > 1:
-        raise QiskitError(f"{name} must be either 0d or 1d.")
-
-    if x_ndim == 1:
-        was_list = True
-    else:
-        x = [x]
-
-    return x, was_list
-
-
-def y0_to_list(y0):
-    """Check if y0 is a single array or list of arrays, and return as a list in either case."""
-    was_list = False
-    if not isinstance(y0, list):
-        y0 = [y0]
-    else:
-        was_list = True
-
-    return y0, was_list
-
-
-def signals_to_list(signals):
-    """Check if signals is a single signal specification or a list of
-    such specifications, and return as a list in either case.
-    """
-    was_list = False
-    if signals is None:
-        signals = [signals]
-    elif isinstance(signals, list) and isinstance(signals[0], (list, SignalList)):
-        # multiple lists
-        was_list = True
-    elif isinstance(signals, SignalList) or (
-        isinstance(signals, list) and not isinstance(signals[0], (list, SignalList))
-    ):
-        # single signals list
-        signals = [signals]
-    else:
-        raise QiskitError("Signals specified in invalid format.")
-
-    return signals, was_list
-
-
-def nested_ndim(x):
-    """Determine the 'ndim' of x, which could be composed of nested lists and array types."""
-    if isinstance(x, (list, tuple)):
-        return 1 + nested_ndim(x[0])
-    elif issubclass(type(x), Dispatch.REGISTERED_TYPES) or isinstance(x, Array):
-        return x.ndim
-
-    # assume scalar
-    return 0
